+++ subgrade-sdk/src/main/java/com/dwarfeng/subgrade/sdk/bean/key/FastJsonLongIdKey.java
@@ -8,8 +8,10 @@ import java.util.Objects;
 
 /**
  * 适用于 FastJson 的 LongIdKey。
- * <p>注意：该类中含有长整型的字段，在与JS前端通信中，会发生精度丢失的问题。
- * 请使用 {@link JSFixedFastJsonLongIdKey} 来解决JS前端的工薪问题</p>
+ *
+ * <p>
+ * 注意：该类中含有长整型的字段，在与JS前端通信中，会发生精度丢失的问题。
+ * 请使用 {@link JSFixedFastJsonLongIdKey} 来解决JS前端的精度丢失问题。
  *
  * @author DwArFeng
  * @since 0.0.1-beta
@@ -29,6 +31,23 @@ public class FastJsonLongIdKey implements Key {
             return null;
         }
         return new FastJsonLongIdKey(longIdKey.getLongId());
+    }
+
+    /**
+     * 根据指定的 FastJsonLongIdKey 生成 LongIdKey。
+     *
+     * @param fastJsonLongIdKey 指定的 FastJsonLongIdKey。
+     * @return 通过指定的 FastJsonLongIdKey 生成的 LongIdKey。
+     * @since 1.2.13
+     */
+    public static LongIdKey toStackBean(FastJsonLongIdKey fastJsonLongIdKey) {
+        if (Objects.isNull(fastJsonLongIdKey)) {
+            return null;
+        } else {
+            return new LongIdKey(
+                    fastJsonLongIdKey.getLongId()
+            );
+        }
     }
 
     @JSONField(name = "long_id", ordinal = 1)